Reflecting comments from Charles Nevile:
[[
The scope section (and probably the charter duration) should point out that the group is expected to be rechartered when the charter expires, or earlier if it takes on deliverables beyond the scope of work identified in the charter.

Web Storage is a Recommendation. We're not planning further work, right?

Gamepad adds the capacity for apps to handle input from typical gaming devices, Pointer Lock allows the app to capture the cursor, and treat it as a motion controller instead of a means of navigation.

I would put service workers under database and offline app apis.

I actually intend to get back to the URL spec. But not for a couple of months.

In Abandoned Work, s/worked/worked on/

We should add the TAG to the liaison groups.
]]

I also included a few other editorial changes.